+++ mods/BrewLAN/units/SEA0002/SEA0002_script.lua
@@ -1,64 +1,24 @@
-local TAirUnit = import('/lua/terranunits.lua').TAirUnit
+local TAirUnit = import('/lua/defaultunits.lua').BrewLANGeoSatelliteUnit
 local VizMarker = import('/lua/sim/VizMarker.lua').VizMarker
 local OffsetBoneToTerrain = import('../../lua/terrainutils.lua').OffsetBoneToTerrain
 
 SEA0002 = Class(TAirUnit) {
     DestroyNoFallRandomChance = 0,
 
-    HideBones = { 'Shell01', 'Shell02', 'Shell03', 'Shell04' },
-
-    OnStopBeingBuilt = function(self, ...)
-        TAirUnit.OnStopBeingBuilt(self, unpack(arg) )
-        --Less fuel for smaller maps. Max fuel only on 81k.
-        self:SetFuelRatio(math.max(ScenarioInfo.size[1], ScenarioInfo.size[2]) / 4096)
-    end,
-
-    PreLaunchSetup = function(self, parent)
+    OnCreate = function(self)
+        TAirUnit.OnCreate(self)
         self:SetScriptBit('RULEUTC_IntelToggle', true)
         self:RemoveToggleCap('RULEUTC_IntelToggle')
     end,
 
-    Setup = function(self, parent)
-        ChangeState( self, self.OpenState )
-    end,
-
-    OpenState = State() {
+    OpenState = State(TAirUnit.OpenState) {
         Main = function(self)
-            self.OpenAnim = CreateAnimator(self)
-            self.OpenAnim:PlayAnim( '/units/XEA0002/xea0002_aopen01.sca' )
-            self.Trash:Add( self.OpenAnim )
-            WaitFor( self.OpenAnim )
-            for k, v in self.HideBones do
-                self:HideBone( v, true )
-            end
-            self.OpenAnim:PlayAnim( '/units/XEA0002/xea0002_aopen02.sca' )
-            WaitFor( self.OpenAnim )
+            TAirUnit.OpenState.Main(self)
             self:CreateVisEntity()
             self:AddToggleCap('RULEUTC_IntelToggle')
             self:SetScriptBit('RULEUTC_IntelToggle', false)
         end,
     },
-
-    StartUnguidedOrbitalDecay = function(self)
-        if not self.UnguidedOrbitalDecay then
-            self.UnguidedOrbitalDecay = self:ForkThread(
-                function()
-                    local pos = self:GetPosition()
-                    while not self.Dead do
-                        self:OnDamage(self, 1, pos, 'Decay')
-                        WaitTicks(1)
-                    end
-                end
-            )
-        end
-    end,
-
-    StopUnguidedOrbitalDecay = function(self)
-        if self.UnguidedOrbitalDecay then
-            KillThread(self.UnguidedOrbitalDecay)
-            self.UnguidedOrbitalDecay = nil
-        end
-    end,
 
     OnIntelEnabled = function(self)
         TAirUnit.OnIntelEnabled(self)
@@ -95,13 +55,6 @@ SEA0002 = Class(TAirUnit) {
         self.VisEntity.Radius = bp.OrbitIntelRadius
         self.Trash:Add(self.VisEntity)
         OffsetBoneToTerrain(self, 'VisEntity')
-    end,
-
-    OnRunOutOfFuel = function(self)
-        TAirUnit.OnRunOutOfFuel(self)
-        self:SetSpeedMult(self:GetBlueprint().Physics.NoFuelSpeedMult)
-        self:SetAccMult(1)
-        self:SetTurnMult(1)
     end,
 }